Added skeleton application bridge
This file allows for you to call different functions by name.  It's
intended that this will be how the Android application and server
application talk to each other.  To play with it, navigate to the file
and request a function.  If the file is located at
127.0.0.1/php/application_bridge.php, then navigate here in your web
browser: 127.0.0.1/php/application_bridge.php?pagr_exec=<function name>